+++ docs/en/changelog.md
@@ -11,6 +11,7 @@ Release version to OpenSTL V0.3.0 as [#25](https://github.com/chengtan9907/OpenS
 * Update STL results visualization by [vis_video](https://github.com/chengtan9907/OpenSTL/tree/master/tools/visualizations/vis_video.py) for video prediction, traffic prediction, weather prediction tasks in [video_visualization](https://github.com/chengtan9907/OpenSTL/docs/en/visualization/video_visualization.md), [traffic_visualization](https://github.com/chengtan9907/OpenSTL/docs/en/visualization/traffic_visualization.md), and [weather_visualization](https://github.com/chengtan9907/SimVPv2/docs/en/visualization/weather_visualization.md).
 * Support Jupyter notebook tutorials and video examples in [examples](https://github.com/chengtan9907/OpenSTL/tree/master/examples).
 * Support early-stop training with `--early_stop_epoch` as issue [#36](https://github.com/chengtan9907/OpenSTL/issues/36).
+* Support inference only with `--inference` in `tools/test.py` for issue [#55](https://github.com/chengtan9907/OpenSTL/issues/55), where results will be saved in `ex_name/saved`.
 
 #### Update Documents